accurately changes categories, updated API routes and front end to reflect all changes
