+++ client/store/index.js
@@ -7,7 +7,8 @@ import user from './user-reducers/user'
 import product from './product-reducers/product'
 import products from './product-reducers/products'
 import search from './search'
-import categories from './category-reducers'
+import categories from './category-reducers/categories'
+import category from './category-reducers/category'
 
 const reducer = combineReducers({
   user,
@@ -15,7 +16,8 @@ const reducer = combineReducers({
   products,
   form: formReducer,
   search,
-  categories
+  categories,
+  category
 })
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
@@ -27,4 +29,5 @@ export * from './user-reducers/user'
 export * from './product-reducers/product'
 export * from './product-reducers/products'
 export * from './search'
-export * from './category-reducers/'
+export * from './category-reducers/categories'
+export * from './category-reducers/category'
